added in role model

+++ src/main/java/com/zhen/mypersonalshop/Service/UserServiceImpl.java
@@ -21,10 +21,6 @@ public class UserServiceImpl implements UserService{
     @Override
     @Transactional
     public User saveUser(User user) {
-        if(user.getCart().size() >= 0)
-        {
-            throw new ResourceAccessException("Cart must be empty");
-        }
         User newUser = new User();
         if(user.getId() != 0)
         {
@@ -33,6 +29,7 @@ public class UserServiceImpl implements UserService{
         newUser.setUsername(user.getUsername().toLowerCase());
         newUser.setEmail(user.getEmail().toLowerCase());
         newUser.setPassword(user.getPassword());
+        return userRepository.save(newUser);
     }
 
     @Override
@@ -57,8 +54,35 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
+    public User updateUser(User user, long id) {
+        Optional<User> activeUser = userRepository.findById(id);
+        if(!activeUser.isPresent())
+        {
+            throw new ResourceNotFoundException("user id " + id + " not found!");
+        }
+        if(user.getUsername() != null)
+        {
+            activeUser.get().setUsername(user.getUsername().toLowerCase());
+        }
+        if(user.getPassword() != null)
+        {
+            activeUser.get().setPassword(user.getPassword());
+        }
+        if(user.getEmail() != null)
+        {
+            activeUser.get().setEmail(user.getEmail().toLowerCase());
+        }
+        return userRepository.save(activeUser.get());
+    }
+
+    @Override
     @Transactional
     public void deleteAccount(long id) {
+        Optional<User> user = userRepository.findById(id);
+        if(!user.isPresent())
+        {
+            throw new ResourceNotFoundException("user id " + id + " not found!");
+        }
         userRepository.deleteById(id);
     }
 }
